fix(purchase): párovací cesta DDKPZ respektuje koš (audit merge)
Sweep koše minul kód § 37a vzniklý paralelně na custom — doklad v koši šlo
napárovat (obousměrně) a odečíst z rekapitulace DPH konečné faktury, zatímco
z výkazů byl vyloučen → podhodnocený odpočet + trvale nesmazatelný doklad.

- PurchaseSettlementService::link()/unlink(): kontrola deleted_at OBOU dokladů
  (protistrana z těla požadavku, kam TrashGuard v akci nedosáhne)
- Link/UnlinkSettlementDoc + DismissAdvanceSuggestion/DismissExtractionWarning:
  TrashGuard::blockIfTrashed po načtení řádku
- settlementDocCandidates/finalCandidates + oba EXISTS flagy ve find():
  AND pi.deleted_at IS NULL (doklad v koši se nenabízí k párování)
- linkAdvance(): deleted_at obou stran (kryje i auto-confirm AI návrhu)
- purchase InvoiceDetail.vue: canMutate = canWrite && !inTrash pro všechny
  mutační prvky detailu
- testy: 4 nové (obousměrné párování, guardy akcí, kandidáti, linkAdvance)

+++ db/migrations/0905_document_trash.sql
@@ -27,7 +27,7 @@ ALTER TABLE purchase_invoices
 
 CREATE TABLE IF NOT EXISTS deleted_document_snapshots (
   id          BIGINT UNSIGNED AUTO_INCREMENT PRIMARY KEY,
-  supplier_id TINYINT UNSIGNED NOT NULL,
+  supplier_id INT UNSIGNED NOT NULL,
   entity_type ENUM('invoice','purchase_invoice') NOT NULL,
   entity_id   BIGINT UNSIGNED NOT NULL,           -- původní id (řádek už neexistuje, bez FK)
   payload     JSON NOT NULL,                      -- kompletní hlavička, položky, úhrady, vazby, cesty k PDF
